requirements.txt
# Core dependencies
numpy>=2.0
opencv-python>=4.8
pillow>=10.0
torch>=2.0

# Video processing
av>=12.0  # PyAV for fast video I/O

# Visualization
matplotlib>=3.8

# SAM 2 (install separately from repo)
# git clone https://github.com/facebookresearch/segment-anything-2.git
# cd segment-anything-2 && pip install -e .

# Optional: supervision for video utilities
# supervision>=0.19
